Implemented 'Doom Flames' using smallengine.
See fabiensanglard.net/doom_fire_psx/index.html for a description. The
code is a mess but is an indication of what features are needed. In
particular, the ability to let the user end the program and eliminating
the need to call SDL_Delay, this should be wrapped in a timer_ function.